+++ example/lib/decoration/barrel_dragable.dart
@@ -5,7 +5,8 @@ import 'package:example/manual_map/dungeon_map.dart';
 import 'package:example/util/common_sprite_sheet.dart';
 import 'package:flutter/material.dart';
 
-class BarrelDraggable extends GameDecoration with DragGesture, ObjectCollision {
+class BarrelDraggable extends GameDecoration
+    with DragGesture, ObjectCollision, TapGesture {
   late TextConfig _textConfig;
   BarrelDraggable(Vector2 position)
       : super.withSprite(
@@ -39,5 +40,25 @@ class BarrelDraggable extends GameDecoration with DragGesture, ObjectCollision {
       'Drag',
       Vector2(this.position.left + width / 5, this.position.top - width / 3),
     );
+  }
+
+  @override
+  void onTap() {
+    print('onTap');
+  }
+
+  @override
+  void onTapCancel() {
+    // TODO: implement onTapCancel
+  }
+
+  @override
+  void onTapDown(int pointer, Offset position) {
+    // TODO: implement onTapDown
+  }
+
+  @override
+  void onTapUp(int pointer, Offset position) {
+    // TODO: implement onTapUp
   }
 }